Eraser, Part 1
Added in a swap layer that is drawn to instead of the background. This
allows the user to draw on top of layers realistically, instead of
drawing on the bottommost layer, as was the case with the previous
implementation. TODO: Fix layer to scan for vertices sent using the
eraser mode and remove them from the list.

Someone needs to do threading >_>